split Viewer and ViewerService
